Extracted ScopeHelper. Unified lifecycle callbacks names. They only operate on AutoScoped Activities and Fragments
