+++ shank2android/src/main/java/life/shank/android/AppContextModule.kt
@@ -1,10 +1,12 @@
 package life.shank.android
 
+import android.annotation.SuppressLint
 import android.content.Context
 import life.shank.ShankModule
 import life.shank.single
 
-object AppContextModule: ShankModule {
+@SuppressLint("StaticFieldLeak")
+object AppContextModule : ShankModule {
     internal lateinit var context: Context
     val appContext = single { -> context }
 }
