+++ shank2android/src/main/java/life/shank/android/AutoScopedFragmentLifecycleCallbacks.kt
@@ -4,17 +4,24 @@ import android.os.Bundle
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
 import life.shank.Scope
-import java.util.*
+import life.shank.android.ScopeHelper.newScope
 
-object ShankFragmentLifecycleListener : FragmentManager.FragmentLifecycleCallbacks() {
-    private val fragmentScopeKey = "shank_fragment_scope_key"
+object AutoScopedFragmentLifecycleCallbacks : FragmentManager.FragmentLifecycleCallbacks() {
+
+    private const val fragmentScopeKey = "shank_fragment_scope_key"
+
     override fun onFragmentPreCreated(fm: FragmentManager, f: Fragment, savedInstanceState: Bundle?) {
-        println("creating fragmeeent")
-        ScopeObservable.putScope(f.hashCode(), savedInstanceState?.getSerializable(fragmentScopeKey) as? Scope
-            ?: Scope(UUID.randomUUID()))
+        if (f !is AutoScoped) return
+
+        ScopeObservable.putScope(
+            f.hashCode(),
+            savedInstanceState?.getSerializable(fragmentScopeKey) as? Scope ?: newScope()
+        )
     }
 
     override fun onFragmentDestroyed(fm: FragmentManager, f: Fragment) {
+        if (f !is AutoScoped) return
+
         if (f.activity?.isFinishing == true || f.isRemoving) {
             ScopeObservable.getScope(f.hashCode()) {
                 it.clear()
@@ -23,8 +30,11 @@ object ShankFragmentLifecycleListener : FragmentManager.FragmentLifecycleCallbac
         ScopeObservable.clear(f.hashCode())
     }
 
-    override fun onFragmentSaveInstanceState(fm: FragmentManager, f: Fragment, outState: Bundle): Unit =
+    override fun onFragmentSaveInstanceState(fm: FragmentManager, f: Fragment, outState: Bundle) {
+        if (f !is AutoScoped) return
+
         ScopeObservable.getScope(f.hashCode()) {
             outState.putSerializable(fragmentScopeKey, it)
         }
+    }
 }
